fix: errors constructed inline to ensure correct stack docs: fence code blocks
